Reindex Elasticsearch wiki index to update integer with long type
Changelog: other
EE: true

+++ ee/lib/elastic/latest/wiki_config.rb
@@ -16,7 +16,7 @@ module Elastic
       mappings dynamic: 'strict' do
         indexes :type, type: :keyword
 
-        indexes :project_id, type: :integer
+        indexes :project_id, type: :long
         indexes :group_id, type: :long
         indexes :rid, type: :keyword
         indexes :oid, type: :keyword, index_options: 'docs', normalizer: :sha_normalizer
